+++ packages/api-lib/libs/es.js
@@ -344,6 +344,38 @@ function buildSort(parameters) {
   return sorting
 }
 
+function buildFieldsFilter(parameters) {
+  const id = 'id'
+  const type = 'type'
+  const bbox = 'bbox'
+  const links = 'links'
+  const assets = 'assets'
+  const { fields } = parameters
+  const _sourceInclude = []
+  const _sourceExclude = []
+  if (fields) {
+    const { geometry, includes, excludes } = fields
+    if (typeof geometry !== 'undefined' && !geometry) {
+      _sourceExclude.push('geometry')
+    }
+    if (includes && includes.length > 0) {
+      const propertiesIncludes = includes.map(
+        (field) => (`properties.${field}`)
+      ).concat(
+        [id, type, bbox, links, assets]
+      )
+      _sourceInclude.push(...propertiesIncludes)
+    }
+    if (excludes && excludes.length > 0) {
+      const filteredExcludes = excludes.filter((field) =>
+        (![id, type, bbox, links, assets].includes(field)))
+      const propertiesExcludes = filteredExcludes.map((field) => (`properties.${field}`))
+      _sourceExclude.push(...propertiesExcludes)
+    }
+  }
+  return { _sourceExclude, _sourceInclude }
+}
+
 async function search(parameters, index = '*', page = 1, limit = 10) {
   let body
   if (parameters.id) {
@@ -354,6 +386,7 @@ async function search(parameters, index = '*', page = 1, limit = 10) {
   }
   const sort = buildSort(parameters)
   body.sort = sort
+
   const searchParams = {
     index,
     body,
@@ -361,6 +394,13 @@ async function search(parameters, index = '*', page = 1, limit = 10) {
     from: (page - 1) * limit
   }
 
+  const { _sourceExclude, _sourceInclude } = buildFieldsFilter(parameters)
+  if (_sourceExclude.length > 0) {
+    searchParams._sourceExclude = _sourceExclude
+  }
+  if (_sourceInclude.length > 0) {
+    searchParams._sourceInclude = _sourceInclude
+  }
   const client = await esClient()
   const resultBody = await client.search(searchParams)
   const results = resultBody.hits.hits.map((r) => (r._source))