+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 - Fix datetime range to be inclusive (i.e., gte and lte)
+- Add `fields` filter to return only specific fields
 
 ## [v0.2.3] - 2019-01-29